Refactor BillingBanner to be more flexible and generic
